backend/requirements.txt
openai
speechrecognition
python-dotenv
google-cloud-texttospeech
simpleaudio